+++ src/collective/dms/basecontent/dmsfile.py
@@ -7,6 +7,7 @@ from plone.autoform import directives as form
 from plone.dexterity.schema import DexteritySchemaPolicy
 from plone.directives.form import default_value
 from Products.CMFPlone.utils import base_hasattr
+from Products.CMFPlone.utils import safe_unicode
 from zope import schema
 from zope.annotation.interfaces import IAnnotations
 from zope.interface import implements
@@ -22,6 +23,8 @@ class IDmsFile(IAnnex):
         title=_(u"Label"),
         required=False,
     )
+
+    form.mode(description="hidden")
 
 
 class DmsFile(Annex):
@@ -51,6 +54,7 @@ class DmsFileSchemaPolicy(DexteritySchemaPolicy):
 class IDmsAppendixFile(IAnnex):
     """Schema for DmsAppendixFile"""
     form.mode(title="hidden")
+    form.mode(description="hidden")
 
 
 class DmsAppendixFile(Annex):
@@ -75,7 +79,7 @@ def titleDefaultValue(data):
         version_number = 1
     else:
         version_number = annotations["higher_version"].value + 1
-    return unicode(version_number)
+    return safe_unicode(version_number)
 
 
 def update_higher_version(context, event):